Deep Learning/Stock Price Prediction with LSTM/requirements.txt
streamlit
tensorflow-cpu
numpy
h5py
